backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
drf-spectacular==0.29.0
pillow==12.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
openai
python-dotenv==1.0.0